Split out the codegen runtime from the reflective lookup
This lets you use the generated code without the reflective runtime (or its ProGuard/R8 rules).
