+++ butterknife-runtime/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=butterknife-runtime
+POM_NAME=ButterKnife Runtime
+POM_PACKAGING=aar
