+++ butterknife-runtime/src/androidTest/java/butterknife/functional/BindBitmapTest.java
@@ -5,12 +5,10 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.support.test.InstrumentationRegistry;
 import butterknife.BindBitmap;
-import butterknife.BindBool;
 import butterknife.Unbinder;
-import butterknife.test.R;
+import butterknife.runtime.test.R;
 import org.junit.Test;
 
-import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertTrue;
 
 public final class BindBitmapTest {
