+++ butterknife-runtime/src/androidTest/java/butterknife/functional/BindDrawableTest.java
@@ -1,18 +1,14 @@
 package butterknife.functional;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.graphics.drawable.Drawable;
 import android.support.test.InstrumentationRegistry;
-import butterknife.BindBitmap;
 import butterknife.BindDrawable;
 import butterknife.Unbinder;
-import butterknife.test.R;
+import butterknife.runtime.test.R;
 import org.junit.Test;
 
 import static com.google.common.truth.Truth.assertThat;
-import static org.junit.Assert.assertTrue;
 
 public final class BindDrawableTest {
   private final Context context = InstrumentationRegistry.getContext();
