+++ butterknife-runtime/src/androidTest/java/butterknife/functional/BindFloatTest.java
@@ -2,25 +2,28 @@ package butterknife.functional;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
-import butterknife.BindString;
+import android.util.TypedValue;
+import butterknife.BindFloat;
 import butterknife.Unbinder;
-import butterknife.test.R;
+import butterknife.runtime.test.R;
 import org.junit.Test;
 
 import static com.google.common.truth.Truth.assertThat;
 
-public final class BindStringTest {
+public final class BindFloatTest {
   private final Context context = InstrumentationRegistry.getContext();
 
   static class Target {
-    @BindString(R.string.hey) String actual;
+    @BindFloat(R.dimen.twelve_point_two) float actual;
   }
 
-  @Test public void simpleInt() {
+  @Test public void asFloat() {
     Target target = new Target();
-    String expected = context.getString(R.string.hey);
+    TypedValue value = new TypedValue();
+    context.getResources().getValue(R.dimen.twelve_point_two, value, true);
+    float expected = value.getFloat();
 
-    Unbinder unbinder = new BindStringTest$Target_ViewBinding(target, context);
+    Unbinder unbinder = new BindFloatTest$Target_ViewBinding(target, context);
     assertThat(target.actual).isEqualTo(expected);
 
     unbinder.unbind();
