+++ butterknife-runtime/src/androidTest/java/butterknife/functional/BindStringTest.java
@@ -2,29 +2,25 @@ package butterknife.functional;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
-import android.util.TypedValue;
-import butterknife.BindFloat;
-import butterknife.BindInt;
+import butterknife.BindString;
 import butterknife.Unbinder;
-import butterknife.test.R;
+import butterknife.runtime.test.R;
 import org.junit.Test;
 
 import static com.google.common.truth.Truth.assertThat;
 
-public final class BindFloatTest {
+public final class BindStringTest {
   private final Context context = InstrumentationRegistry.getContext();
 
   static class Target {
-    @BindFloat(R.dimen.twelve_point_two) float actual;
+    @BindString(R.string.hey) String actual;
   }
 
-  @Test public void asFloat() {
+  @Test public void simpleInt() {
     Target target = new Target();
-    TypedValue value = new TypedValue();
-    context.getResources().getValue(R.dimen.twelve_point_two, value, true);
-    float expected = value.getFloat();
+    String expected = context.getString(R.string.hey);
 
-    Unbinder unbinder = new BindFloatTest$Target_ViewBinding(target, context);
+    Unbinder unbinder = new BindStringTest$Target_ViewBinding(target, context);
     assertThat(target.actual).isEqualTo(expected);
 
     unbinder.unbind();
