+++ butterknife-runtime/src/androidTest/java/butterknife/internal/UtilsTest.java
