+++ butterknife-runtime/src/androidTest/res/color/colors.xml
