+++ butterknife-runtime/src/androidTest/res/drawable/circle.xml
