+++ butterknife-runtime/src/androidTest/res/drawable/pixel.png
