+++ butterknife-runtime/src/androidTest/res/values/values.xml
