+++ butterknife-runtime/src/test/java/butterknife/BindAnimTest.java
