+++ butterknife-runtime/src/test/java/butterknife/BindBitmapTest.java
