+++ butterknife-runtime/src/test/java/butterknife/BindBoolTest.java
