+++ butterknife-runtime/src/test/java/butterknife/BindColorTest.java
