+++ butterknife-runtime/src/test/java/butterknife/BindDrawableTest.java
