+++ butterknife-runtime/src/test/java/butterknife/BindFloatTest.java
