+++ butterknife-runtime/src/test/java/butterknife/BindStringTest.java
