+++ butterknife-runtime/src/test/java/butterknife/BindViewsTest.java
