+++ butterknife-runtime/src/test/java/butterknife/ExtendActivityTest.java
