+++ butterknife-runtime/src/test/java/butterknife/ExtendDialogTest.java
