+++ butterknife/build.gradle
@@ -9,12 +9,6 @@ android {
     consumerProguardFiles 'proguard-rules.txt'
 
     testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
-
-    javaCompileOptions {
-      annotationProcessorOptions {
-        includeCompileClasspath = true
-      }
-    }
   }
 
   lintOptions {
@@ -31,36 +25,12 @@ android {
 }
 
 dependencies {
-  api project(':butterknife-annotations')
-  api deps.support.compat
-
-  lintChecks project(':butterknife-lint')
+  api project(':butterknife-runtime')
 
   androidTestImplementation deps.junit
   androidTestImplementation deps.truth
   androidTestImplementation deps.support.test.runner
   androidTestAnnotationProcessor project(':butterknife-compiler')
-
-  testImplementation deps.junit
-  testImplementation deps.truth
-  testImplementation deps.compiletesting
-  testImplementation files(getRuntimeJar())
-  testImplementation files(org.gradle.internal.jvm.Jvm.current().getToolsJar())
-  testImplementation project(':butterknife-compiler')
-}
-
-def getRuntimeJar() {
-  try {
-    final File javaBase = new File(System.getProperty("java.home")).getCanonicalFile();
-    File runtimeJar = new File(javaBase, "lib/rt.jar");
-    if (runtimeJar.exists()) {
-      return runtimeJar;
-    }
-    runtimeJar = new File(javaBase, "jre/lib/rt.jar");
-    return runtimeJar.exists() ? runtimeJar : null;
-  } catch (IOException e) {
-    throw new RuntimeException(e);
-  }
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
