+++ settings.gradle
@@ -2,8 +2,9 @@ include ':butterknife'
 include ':butterknife-annotations'
 include ':butterknife-compiler'
 include ':butterknife-gradle-plugin'
-include ':butterknife-lint'
 include ':butterknife-integration-test'
+include ':butterknife-lint'
+include ':butterknife-runtime'
 
 //include ':sample:app'
 //include ':sample:library'
